+++ D35E.js
@@ -133,7 +133,7 @@ Hooks.once("setup", function() {
  * Once the entire VTT framework is initialized, check to see if we should perform a data migration
  */
 Hooks.once("ready", async function() {
-  const NEEDS_MIGRATION_VERSION = "0.84.0";
+  const NEEDS_MIGRATION_VERSION = "0.85.0";
   let PREVIOUS_MIGRATION_VERSION = game.settings.get("D35E", "systemMigrationVersion");
   if (typeof PREVIOUS_MIGRATION_VERSION === "number") {
     PREVIOUS_MIGRATION_VERSION = PREVIOUS_MIGRATION_VERSION.toString() + ".0";
@@ -157,8 +157,14 @@ Hooks.once("ready", async function() {
     TopPortraitBar.render(game.actors.get(key))
   }
 
-  if (!game.user.isGM)
+  if (!game.user.isGM) {
+    (await import(
+            /* webpackChunkName: "welcome-screen" */
+            './module/onboarding.js'
+            )
+    ).default();
     return;
+  }
   // Edit next line to match module.
   const system = game.system;
   const title = system.data.title;
@@ -173,7 +179,11 @@ Hooks.once("ready", async function() {
 
   if (!isNewerVersion(moduleVersion, oldVersion))
     return;
-
+  (await import(
+          /* webpackChunkName: "welcome-screen" */
+          './module/onboarding.js'
+          )
+  ).default();
   (await import(
           /* webpackChunkName: "welcome-screen" */
           './module/welcome-screen.js'