+++ module/settings.js
@@ -240,4 +240,13 @@ export const registerSystemSettings = function() {
     // type: Boolean,
   // });
 
+
+  game.settings.register("D35E", '__onboarding', {
+    name: `Tutorial shown`,
+    hint: 'Basic system usage tutorial already shown. Uncheck to view again after reload.',
+    default: false,
+    type: Boolean,
+    config: true,
+    scope: 'client',
+  });
 };